:truck: move `internal/logging` to `internal/app/logging`
