+++ cmd/flowg/serve.go
@@ -14,8 +14,8 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"link-society.com/flowg/internal/app/logging"
 	"link-society.com/flowg/internal/auth"
-	"link-society.com/flowg/internal/logging"
 	"link-society.com/flowg/internal/logstorage"
 	"link-society.com/flowg/internal/pipelines"
