+++ internal/app/logging/badger.go
